[cache] make sure `get_metadata` doesn't crash if HDF is invalid
Pandas does not always close the HDF store when it's invalid. Since we catch
all exception to write an empty metadata file, it crashes because the HDF file
is still opened. cf. https://github.com/pandas-dev/pandas/issues/28430

If we handle the file closing ourselves and pass the file-object to `pd.read_hdf`,
everything should be fine BUT unfortunetaly `pd.read_hdf` does not actually
support buffer: "Support for generic buffers has not been implemented."

So let's instantiate HDFStore directly to be able to close it no matter what