fix(excel): interpret null values correctly
Not as empty strings
Now that we're transforming excel into CSV first, we need a way to
encode null values across this process.
Before, the `keep_default_na` were falsy by default, so no values was
ever considered null. We changed that so we use the default list that
pandas provide.
As a consequence, call that contains 'N/A', 'null', etc. that were read
as strings are now read as None/NaN.