fix: default timeout was originally at 60s,
but for each ftp-client it was defined at 3s,
Setting it to 10s is too small when we have a long list of files that could takes too much time to resolve,
the value should be increased, in this commit, am tripling the default fomr FTPS class that was at 60s to 180s.

Later on, we should find a way to set a FF that will forward that
extremly HUGE value to the lib itself.